Split "FeatureLayer"s into Symbolizers, add Leafdoc comments all around, and add OpenMapTiles to demo.
