+++ demo/demo-vectortiles.html
@@ -6,12 +6,12 @@
 
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-	<link rel="stylesheet" href="https://unpkg.com/leaflet@1.0.2/dist/leaflet.css" />
-	<script src="https://unpkg.com/leaflet@1.0.2/dist/leaflet.js"></script>
+	<link rel="stylesheet" href="https://unpkg.com/leaflet@1.0.3/dist/leaflet.css" />
+	<script src="https://unpkg.com/leaflet@1.0.3/dist/leaflet.js"></script>
 	<script src="../dist/Leaflet.VectorGrid.bundled.js"></script>
 </head>
 <body style='margin:0'>
-	<div id="map" style="width: 100vw; height: 100vh"></div>
+	<div id="map" style="width: 100vw; height: 100vh; background: white"></div>
 
 
 <!--	<script type="text/javascript" src="../vendor/pbf-dev.js"></script>
@@ -34,61 +34,189 @@
 // 		}).addTo(map);
 
 
-		var url = 'https://{s}.tiles.mapbox.com/v4/mapbox.mapbox-streets-v6/{z}/{x}/{y}.vector.pbf?access_token=pk.eyJ1IjoibWFwYm94IiwiYSI6ImNpandmbXliNDBjZWd2M2x6bDk3c2ZtOTkifQ._QA7i5Mpkd_m30IGElHziw';
+		var vectorTileStyling = {
 
-		var vectorTileOptions = {
-			rendererFactory: L.canvas.tile,
-			attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, &copy; <a href="https://www.mapbox.com/about/maps/">MapBox</a>',
-			vectorTileLayerStyles: {
-
-				water: {
-					weight: 0,
-					fillColor: '#9bc2c4',
-					fillOpacity: 1,
-					fill: true,
-					stroke: false
-				},
-
-				admin: function(properties, zoom) {
-					var level = properties.admin_level;
-					var weight = 1;
-					if (level == 2) {weight = 4;}
-					return {
-						weight: weight,
-						color: '#cf52d3',
-						dashArray: '2, 6',
-						fill: false,
-// 						lineCap: 'butt',
-						lineJoin: 'miter'
-					}
-				},
-
-				country_label: [],
-				marine_label: [],
-				state_label: [],
-				place_label: [],
-				waterway_label: [],
-				landuse: [],
-				landuse_overlay: [],
-				road: [],
-				poi_label: [],
-				waterway: [],
-				aeroway: [],
-				tunnel: [],
-				bridge: [],
-				barrier_line: [],
-				building: [],
-				road_label: [],
-				housenum_label: [],
-
+			water: {
+				fill: true,
+				weight: 1,
+				fillColor: '#06cccc',
+				color: '#06cccc',
+				fillOpacity: 0.2,
+				opacity: 0.4,
 			},
+			admin: {
+				weight: 1,
+				fillColor: 'pink',
+				color: 'pink',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			waterway: {
+				weight: 1,
+				fillColor: '#2375e0',
+				color: '#2375e0',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			landcover: {
+				fill: true,
+				weight: 1,
+				fillColor: '#53e033',
+				color: '#53e033',
+				fillOpacity: 0.2,
+				opacity: 0.4,
+			},
+			landuse: {
+				fill: true,
+				weight: 1,
+				fillColor: '#e5b404',
+				color: '#e5b404',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			park: {
+				fill: true,
+				weight: 1,
+				fillColor: '#84ea5b',
+				color: '#84ea5b',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			boundary: {
+				weight: 1,
+				fillColor: '#c545d3',
+				color: '#c545d3',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			aeroway: {
+				weight: 1,
+				fillColor: '#51aeb5',
+				color: '#51aeb5',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			road: {
+				weight: 1,
+				fillColor: '#f2b648',
+				color: '#f2b648',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			tunnel: {
+				weight: 0.5,
+				fillColor: '#f2b648',
+				color: '#f2b648',
+				fillOpacity: 0.2,
+				opacity: 0.4,
+// 					dashArray: [4, 4]
+			},
+			bridge: {
+				weight: 0.5,
+				fillColor: '#f2b648',
+				color: '#f2b648',
+				fillOpacity: 0.2,
+				opacity: 0.4,
+// 					dashArray: [4, 4]
+			},
+			building: {
+				fill: true,
+				weight: 1,
+				fillColor: '#2b2b2b',
+				color: '#2b2b2b',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			water_name: {
+				weight: 1,
+				fillColor: '#022c5b',
+				color: '#022c5b',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			transportation_name: {
+				weight: 1,
+				fillColor: '#bc6b38',
+				color: '#bc6b38',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			place: {
+				weight: 1,
+				fillColor: '#f20e93',
+				color: '#f20e93',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			housenumber: {
+				weight: 1,
+				fillColor: '#ef4c8b',
+				color: '#ef4c8b',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+			poi: {
+				weight: 1,
+				fillColor: '#3bb50a',
+				color: '#3bb50a',
+				fillOpacity: 0.2,
+				opacity: 0.4
+			},
+
+
+			country_label: [],
+			marine_label: [],
+			state_label: [],
+			place_label: [],
+			waterway_label: [],
+// 				landuse: [],
+			landuse_overlay: [],
+// 				road: [],
+			poi_label: [],
+// 				waterway: [],
+// 				aeroway: [],
+// 				tunnel: [],
+// 				bridge: [],
+			barrier_line: [],
+// 				building: [],
+			road_label: [],
+			housenum_label: [],
 		};
 
 
-		var pbfLayer = L.vectorGrid.protobuf(url, vectorTileOptions).addTo(map);
+
+
+		var mapboxUrl = 'https://{s}.tiles.mapbox.com/v4/mapbox.mapbox-streets-v6/{z}/{x}/{y}.vector.pbf?access_token=pk.eyJ1IjoibWFwYm94IiwiYSI6ImNpandmbXliNDBjZWd2M2x6bDk3c2ZtOTkifQ._QA7i5Mpkd_m30IGElHziw';
+
+		var mapboxVectorTileOptions = {
+			rendererFactory: L.canvas.tile,
+			attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, &copy; <a href="https://www.mapbox.com/about/maps/">MapBox</a>',
+			vectorTileLayerStyles: vectorTileStyling,
+		};
+
+
+		var mapboxPbfLayer = L.vectorGrid.protobuf(mapboxUrl, mapboxVectorTileOptions);
+
+
+
+		var openmaptilesUrl = 'https://{s}.tiles.mapbox.com/v4/mapbox.mapbox-streets-v6/{z}/{x}/{y}.vector.pbf?access_token=pk.eyJ1IjoibWFwYm94IiwiYSI6ImNpandmbXliNDBjZWd2M2x6bDk3c2ZtOTkifQ._QA7i5Mpkd_m30IGElHziw';
+
+		var openmaptilesVectorTileOptions = {
+			rendererFactory: L.canvas.tile,
+			attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors, &copy; <a href="https://www.mapbox.com/about/maps/">MapBox</a>',
+			vectorTileLayerStyles: vectorTileStyling,
+		};
+
+
+		var openmaptilesPbfLayer = L.vectorGrid.protobuf(openmaptilesUrl, openmaptilesVectorTileOptions).addTo(map);
 
 		map.setView({ lat: 47.040182144806664, lng: 9.667968750000002 }, 0);
 
+
+		L.control.layers({
+			OpenMapTiles: openmaptilesPbfLayer,
+			"MapBox Vector Tiles": mapboxPbfLayer
+		}, {}, {collapsed: false}).addTo(map);
 
 	</script>
 </body>
